feat: support language aliases for code editor (#9336)
* feat(CodeEditor): support language aliases

* fix: mermaid

* refactor: lookup

* chore: sort package.json
