+++ src/renderer/src/components/CodeBlockView/view.tsx
@@ -18,8 +18,8 @@ import { BasicPreviewHandles } from '@renderer/components/Preview'
 import { MAX_COLLAPSED_CODE_HEIGHT } from '@renderer/config/constant'
 import { useSettings } from '@renderer/hooks/useSettings'
 import { pyodideService } from '@renderer/services/PyodideService'
+import { getExtensionByLanguage } from '@renderer/utils/code-language'
 import { extractTitle } from '@renderer/utils/formats'
-import { getExtensionByLanguage } from '@renderer/utils/markdown'
 import dayjs from 'dayjs'
 import React, { memo, startTransition, useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { useTranslation } from 'react-i18next'
