+++ src/renderer/src/context/CodeStyleProvider.tsx
@@ -20,7 +20,6 @@ interface CodeStyleContextType {
   activeShikiTheme: string
   isShikiThemeDark: boolean
   activeCmTheme: any
-  languageMap: Record<string, string>
 }
 
 const defaultCodeStyleContext: CodeStyleContextType = {
@@ -33,8 +32,7 @@ const defaultCodeStyleContext: CodeStyleContextType = {
   themeNames: ['auto'],
   activeShikiTheme: 'auto',
   isShikiThemeDark: false,
-  activeCmTheme: null,
-  languageMap: {}
+  activeCmTheme: null
 }
 
 const CodeStyleContext = createContext<CodeStyleContextType>(defaultCodeStyleContext)
@@ -93,8 +91,8 @@ export const CodeStyleProvider: React.FC<PropsWithChildren> = ({ children }) => 
     return cmThemes[themeName as keyof typeof cmThemes] || themeName
   }, [theme, codeEditor, themeNames])
 
-  // 一些语言的别名
-  const languageMap = useMemo(() => {
+  // 自定义 shiki 语言别名
+  const languageAliases = useMemo(() => {
     return {
       bash: 'shell',
       'objective-c++': 'objective-cpp',
@@ -114,10 +112,10 @@ export const CodeStyleProvider: React.FC<PropsWithChildren> = ({ children }) => 
   // 流式代码高亮，返回已高亮的 token lines
   const highlightCodeChunk = useCallback(
     async (trunk: string, language: string, callerId: string) => {
-      const normalizedLang = languageMap[language as keyof typeof languageMap] || language.toLowerCase()
+      const normalizedLang = languageAliases[language as keyof typeof languageAliases] || language.toLowerCase()
       return shikiStreamService.highlightCodeChunk(trunk, normalizedLang, activeShikiTheme, callerId)
     },
-    [activeShikiTheme, languageMap]
+    [activeShikiTheme, languageAliases]
   )
 
   // 清理代码高亮资源
@@ -128,19 +126,19 @@ export const CodeStyleProvider: React.FC<PropsWithChildren> = ({ children }) => 
   // 高亮流式输出的代码
   const highlightStreamingCode = useCallback(
     async (fullContent: string, language: string, callerId: string) => {
-      const normalizedLang = languageMap[language as keyof typeof languageMap] || language.toLowerCase()
+      const normalizedLang = languageAliases[language as keyof typeof languageAliases] || language.toLowerCase()
       return shikiStreamService.highlightStreamingCode(fullContent, normalizedLang, activeShikiTheme, callerId)
     },
-    [activeShikiTheme, languageMap]
+    [activeShikiTheme, languageAliases]
   )
 
   // 获取 Shiki pre 标签属性
   const getShikiPreProperties = useCallback(
     async (language: string) => {
-      const normalizedLang = languageMap[language as keyof typeof languageMap] || language.toLowerCase()
+      const normalizedLang = languageAliases[language as keyof typeof languageAliases] || language.toLowerCase()
       return shikiStreamService.getShikiPreProperties(normalizedLang, activeShikiTheme)
     },
-    [activeShikiTheme, languageMap]
+    [activeShikiTheme, languageAliases]
   )
 
   const highlightCode = useCallback(
@@ -176,8 +174,7 @@ export const CodeStyleProvider: React.FC<PropsWithChildren> = ({ children }) => 
       themeNames,
       activeShikiTheme,
       isShikiThemeDark,
-      activeCmTheme,
-      languageMap
+      activeCmTheme
     }),
     [
       highlightCodeChunk,
@@ -189,8 +186,7 @@ export const CodeStyleProvider: React.FC<PropsWithChildren> = ({ children }) => 
       themeNames,
       activeShikiTheme,
       isShikiThemeDark,
-      activeCmTheme,
-      languageMap
+      activeCmTheme
     ]
   )
